Added all remaining required functions and completed the assignment.
playRound calls upon getComputerChoice and getPlayerChoice as parameters to compare their values to complete a round of RPS.
playGame calls upon playRound 5 times to complete a 3 out of 5 set of RPS and displays the winners of each round and the entire match.

This was a lot of fun, thanks guys.